Delete control

+++ app/Http/Controllers/AdminCampaignController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Campaign;
+use App\Models\Product;
 use Illuminate\Http\Request;
 
 class AdminCampaignController extends Controller
@@ -44,7 +45,14 @@ class AdminCampaignController extends Controller
 
     public function destroy(Request $request)
     {
-        $campaign = Campaign::query()->where('id',$request->id)->delete();
+        //verificamos que no existe ningún producto en esa campaña.
+        $product=Product::whereNotNull('campaign')->where('campaign',$request->id)->first();
+        if($product) {
+            $campaign = [ "error" => __("There are products in this Campaign. It is not possible to delete it.") ];
+            //alert()->error(__('Whoops!'),__('There are products in this Campaign. It is not possible to delete it.'));
+        } else {
+            $campaign = Campaign::query()->where('id',$request->id)->delete();
+        }
         return Response()->json($campaign);
     }
